# 2.0.0

* [CHORE] drop support for Python 2.7, 3.5 and 3.6
* [CHORE] migrate pytest.ini configuration into pyproject.toml
* [CHORE] migrate from setuptools to hatch / hatchling
* [CHORE] add types
* [CHORE] add types to tests
* [CHORE] update dependencies
* [CHORE] use f-strings where appropriate

+++ tests/conftest.py
@@ -1,32 +1,38 @@
 import json
 import socket
+import typing as t
 from codecs import open
-from collections import namedtuple
-from contextlib import closing, contextmanager
+from contextlib import contextmanager
 from os.path import abspath, dirname, isfile, join, realpath
+from pathlib import Path
 from time import sleep
 
 import docker
 import mysql.connector
 import pytest
-import six
+from _pytest._py.path import LocalPath
+from _pytest.config import Config
+from _pytest.config.argparsing import Parser
+from _pytest.legacypath import TempdirFactory
 from click.testing import CliRunner
+from docker import DockerClient
 from docker.errors import NotFound
-from mysql.connector import errorcode
+from docker.models.containers import Container
+from faker import Faker
+from mysql.connector import CMySQLConnection, MySQLConnection, errorcode
+from mysql.connector.pooling import PooledMySQLConnection
 from requests import HTTPError
 from sqlalchemy import create_engine
+from sqlalchemy.engine import Engine
 from sqlalchemy.exc import IntegrityError
+from sqlalchemy.orm import Session
 from sqlalchemy_utils import database_exists, drop_database
 
 from .database import Database
 from .factories import ArticleFactory, AuthorFactory, ImageFactory, MediaFactory, MiscFactory, TagFactory
 
 
-if six.PY2:
-    from sixeptions import *
-
-
-def pytest_addoption(parser):
+def pytest_addoption(parser: "Parser") -> None:
     parser.addoption(
         "--sqlite-file",
         dest="sqlite_file",
@@ -89,9 +95,9 @@ def pytest_addoption(parser):
 
 
 @pytest.fixture(scope="session", autouse=True)
-def cleanup_hanged_docker_containers():
+def cleanup_hanged_docker_containers() -> None:
     try:
-        client = docker.from_env()
+        client: DockerClient = docker.from_env()
         for container in client.containers.list():
             if container.name == "pytest_sqlite3_to_mysql":
                 container.kill()
@@ -100,9 +106,9 @@ def cleanup_hanged_docker_containers():
         pass
 
 
-def pytest_keyboard_interrupt():
+def pytest_keyboard_interrupt() -> None:
     try:
-        client = docker.from_env()
+        client: DockerClient = docker.from_env()
         for container in client.containers.list():
             if container.name == "pytest_sqlite3_to_mysql":
                 container.kill()
@@ -114,17 +120,17 @@ def pytest_keyboard_interrupt():
 class Helpers:
     @staticmethod
     @contextmanager
-    def not_raises(exception):
+    def not_raises(exception: t.Type[Exception]) -> t.Generator:
         try:
             yield
         except exception:
-            raise pytest.fail("DID RAISE {0}".format(exception))
+            raise pytest.fail(f"DID RAISE {exception}")
 
     @staticmethod
     @contextmanager
-    def session_scope(db):
+    def session_scope(db: Database) -> t.Generator:
         """Provide a transactional scope around a series of operations."""
-        session = db.Session()
+        session: Session = db.Session()
         try:
             yield session
             session.commit()
@@ -136,114 +142,72 @@ class Helpers:
 
 
 @pytest.fixture
-def helpers():
+def helpers() -> t.Type[Helpers]:
     return Helpers
 
 
-if six.PY2:
+@pytest.fixture(scope="session")
+def sqlite_database(pytestconfig: Config, _session_faker: Faker, tmpdir_factory: TempdirFactory) -> str:
+    db_file: LocalPath = pytestconfig.getoption("sqlite_file")
+    if db_file:
+        if not isfile(realpath(db_file)):
+            pytest.fail(f"{db_file} does not exist")
+        return str(realpath(db_file))
 
-    @pytest.fixture(scope="session")
-    def sqlite_database(pytestconfig, faker, tmpdir_factory):
-        db_file = pytestconfig.getoption("sqlite_file")
-        if db_file:
-            if not isfile(realpath(db_file)):
-                pytest.fail("{} does not exist".format(db_file))
-                raise FileNotFoundError("{} does not exist".format(db_file))
-            return realpath(db_file)
+    temp_data_dir: LocalPath = tmpdir_factory.mktemp("data")
+    temp_image_dir: LocalPath = tmpdir_factory.mktemp("images")
+    db_file = temp_data_dir.join(Path("db.sqlite3"))
+    db: Database = Database(f"sqlite:///{db_file}")
 
-        temp_data_dir = tmpdir_factory.mktemp("data")
-        temp_image_dir = tmpdir_factory.mktemp("images")
-        db_file = temp_data_dir.join("db.sqlite3")
-        db = Database("sqlite:///{}".format(str(db_file)))
-
-        with Helpers.session_scope(db) as session:
-            for _ in range(faker.pyint(min_value=12, max_value=24)):
-                article = ArticleFactory()
-                article.authors.append(AuthorFactory())
-                article.tags.append(TagFactory())
-                article.misc.append(MiscFactory())
-                article.media.append(MediaFactory())
-                for _ in range(faker.pyint(min_value=1, max_value=4)):
-                    article.images.append(
-                        ImageFactory(
-                            path=join(
-                                str(temp_image_dir),
-                                faker.year(),
-                                faker.month(),
-                                faker.day_of_month(),
-                                faker.file_name(extension="jpg"),
-                            )
+    with Helpers.session_scope(db) as session:
+        for _ in range(_session_faker.pyint(min_value=12, max_value=24)):
+            article = ArticleFactory()
+            article.authors.append(AuthorFactory())
+            article.tags.append(TagFactory())
+            article.misc.append(MiscFactory())
+            article.media.append(MediaFactory())
+            for _ in range(_session_faker.pyint(min_value=1, max_value=4)):
+                article.images.append(
+                    ImageFactory(
+                        path=join(
+                            str(temp_image_dir),
+                            _session_faker.year(),
+                            _session_faker.month(),
+                            _session_faker.day_of_month(),
+                            _session_faker.file_name(extension="jpg"),
                         )
                     )
-                session.add(article)
-            try:
-                session.commit()
-            except IntegrityError:
-                session.rollback()
+                )
+            session.add(article)
+        try:
+            session.commit()
+        except IntegrityError:
+            session.rollback()
 
-        return str(db_file)
-
-else:
-
-    @pytest.fixture(scope="session")
-    def sqlite_database(pytestconfig, _session_faker, tmpdir_factory):
-        db_file = pytestconfig.getoption("sqlite_file")
-        if db_file:
-            if not isfile(realpath(db_file)):
-                pytest.fail("{} does not exist".format(db_file))
-                raise FileNotFoundError("{} does not exist".format(db_file))
-            return realpath(db_file)
-
-        temp_data_dir = tmpdir_factory.mktemp("data")
-        temp_image_dir = tmpdir_factory.mktemp("images")
-        db_file = temp_data_dir.join("db.sqlite3")
-        db = Database("sqlite:///{}".format(str(db_file)))
-
-        with Helpers.session_scope(db) as session:
-            for _ in range(_session_faker.pyint(min_value=12, max_value=24)):
-                article = ArticleFactory()
-                article.authors.append(AuthorFactory())
-                article.tags.append(TagFactory())
-                article.misc.append(MiscFactory())
-                article.media.append(MediaFactory())
-                for _ in range(_session_faker.pyint(min_value=1, max_value=4)):
-                    article.images.append(
-                        ImageFactory(
-                            path=join(
-                                str(temp_image_dir),
-                                _session_faker.year(),
-                                _session_faker.month(),
-                                _session_faker.day_of_month(),
-                                _session_faker.file_name(extension="jpg"),
-                            )
-                        )
-                    )
-                session.add(article)
-            try:
-                session.commit()
-            except IntegrityError:
-                session.rollback()
-
-        return str(db_file)
+    return str(db_file)
 
 
-def is_port_in_use(port, host="0.0.0.0"):
-    if six.PY2:
-        with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
-            return s.connect_ex((host, port)) == 0
-    else:
-        with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-            return s.connect_ex((host, port)) == 0
+def is_port_in_use(port: int, host: str = "0.0.0.0") -> bool:
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        return s.connect_ex((host, port)) == 0
+
+
+class MySQLCredentials(t.NamedTuple):
+    """MySQL credentials."""
+
+    user: str
+    password: str
+    host: str
+    port: int
+    database: str
 
 
 @pytest.fixture(scope="session")
-def mysql_credentials(pytestconfig):
-    MySQLCredentials = namedtuple("MySQLCredentials", ["user", "password", "host", "port", "database"])
-
-    db_credentials_file = abspath(join(dirname(__file__), "db_credentials.json"))
+def mysql_credentials(pytestconfig: Config) -> MySQLCredentials:
+    db_credentials_file: str = abspath(join(dirname(__file__), "db_credentials.json"))
     if isfile(db_credentials_file):
         with open(db_credentials_file, "r", "utf-8") as fh:
-            db_credentials = json.load(fh)
+            db_credentials: t.Dict[str, t.Any] = json.load(fh)
             return MySQLCredentials(
                 user=db_credentials["mysql_user"],
                 password=db_credentials["mysql_password"],
@@ -252,16 +216,11 @@ def mysql_credentials(pytestconfig):
                 port=db_credentials["mysql_port"],
             )
 
-    port = pytestconfig.getoption("mysql_port") or 3306
+    port: int = pytestconfig.getoption("mysql_port") or 3306
     if pytestconfig.getoption("use_docker"):
         while is_port_in_use(port, pytestconfig.getoption("mysql_host")):
             if port >= 2**16 - 1:
-                pytest.fail(
-                    "No ports appear to be available on the host {}".format(pytestconfig.getoption("mysql_host"))
-                )
-                raise ConnectionError(
-                    "No ports appear to be available on the host {}".format(pytestconfig.getoption("mysql_host"))
-                )
+                pytest.fail(f'No ports appear to be available on the host {pytestconfig.getoption("mysql_host")}')
             port += 1
 
     return MySQLCredentials(
@@ -274,11 +233,11 @@ def mysql_credentials(pytestconfig):
 
 
 @pytest.fixture(scope="session")
-def mysql_instance(mysql_credentials, pytestconfig):
-    container = None
-    mysql_connection = None
-    mysql_available = False
-    mysql_connection_retries = 15  # failsafe
+def mysql_instance(mysql_credentials: MySQLCredentials, pytestconfig: Config) -> t.Iterator[MySQLConnection]:
+    container: t.Optional[Container] = None
+    mysql_connection: t.Optional[t.Union[PooledMySQLConnection, MySQLConnection, CMySQLConnection]] = None
+    mysql_available: bool = False
+    mysql_connection_retries: int = 15  # failsafe
 
     db_credentials_file = abspath(join(dirname(__file__), "db_credentials.json"))
     if isfile(db_credentials_file):
@@ -294,17 +253,15 @@ def mysql_instance(mysql_credentials, pytestconfig):
             client = docker.from_env()
         except Exception as err:
             pytest.fail(str(err))
-            raise
 
         docker_mysql_image = pytestconfig.getoption("docker_mysql_image") or "mysql:latest"
 
         if not any(docker_mysql_image in image.tags for image in client.images.list()):
-            print("Attempting to download Docker image {}'".format(docker_mysql_image))
+            print(f"Attempting to download Docker image {docker_mysql_image}'")
             try:
                 client.images.pull(docker_mysql_image)
             except (HTTPError, NotFound) as err:
                 pytest.fail(str(err))
-                raise
 
         container = client.containers.run(
             image=docker_mysql_image,
@@ -312,7 +269,7 @@ def mysql_instance(mysql_credentials, pytestconfig):
             ports={
                 "3306/tcp": (
                     mysql_credentials.host,
-                    "{}/tcp".format(mysql_credentials.port),
+                    f"{mysql_credentials.port}/tcp",
                 )
             },
             environment={
@@ -352,24 +309,18 @@ def mysql_instance(mysql_credentials, pytestconfig):
         if not mysql_available and mysql_connection_retries <= 0:
             raise ConnectionAbortedError("Maximum MySQL connection retries exhausted! Are you sure MySQL is running?")
 
-    yield
+    yield  # type: ignore[misc]
 
-    if use_docker:
+    if use_docker and container is not None:
         container.kill()
 
 
 @pytest.fixture()
-def mysql_database(mysql_instance, mysql_credentials):
-    yield
+def mysql_database(mysql_instance: t.Generator, mysql_credentials: MySQLCredentials) -> t.Iterator[Engine]:
+    yield  # type: ignore[misc]
 
-    engine = create_engine(
-        "mysql+pymysql://{user}:{password}@{host}:{port}/{database}".format(
-            user=mysql_credentials.user,
-            password=mysql_credentials.password,
-            host=mysql_credentials.host,
-            port=mysql_credentials.port,
-            database=mysql_credentials.database,
-        )
+    engine: Engine = create_engine(
+        f"mysql+pymysql://{mysql_credentials.user}:{mysql_credentials.password}@{mysql_credentials.host}:{mysql_credentials.port}/{mysql_credentials.database}"
     )
 
     if database_exists(engine.url):
@@ -377,5 +328,5 @@ def mysql_database(mysql_instance, mysql_credentials):
 
 
 @pytest.fixture()
-def cli_runner():
+def cli_runner() -> t.Iterator[CliRunner]:
     yield CliRunner()
